Fix run-outcome recording and guard against tight re-run loops
recordRun bound $7 to both last_new_record_count (integer) and
total_records (bigint), so Postgres refused the statement with
'inconsistent types deduced for parameter $7'. Every run therefore
fetched and stored records correctly but failed to record its outcome,
leaving last_run_at NULL — which made the monitor look due on every
30s tick instead of hourly. Numeric parameters are now cast explicitly.

The scheduler also tracks in-process attempt times, so a failure in the
outcome-recording path can no longer turn into a tight loop against an
upstream source.

+++ src/store/registry.ts
@@ -136,13 +136,15 @@ export async function recordRun(client: PoolClient, run: RunResult): Promise<Mon
        last_status           = $3,
        last_error            = $4,
        last_success_at       = case when $5 then $2 else last_success_at end,
-       last_record_count     = $6,
-       last_new_record_count = $7,
-       last_duration_ms      = $8,
+       last_record_count     = $6::integer,
+       last_new_record_count = $7::integer,
+       last_duration_ms      = $8::integer,
        consecutive_failures  = case when $5 then 0 else consecutive_failures + 1 end,
        total_runs            = total_runs + 1,
        total_failures        = total_failures + case when $5 then 0 else 1 end,
-       total_records         = total_records + $7,
+       -- $7 also feeds a bigint column; without the cast Postgres cannot deduce
+       -- a single type for the parameter and rejects the whole statement.
+       total_records         = total_records + $7::integer,
        stale_alert_at        = case when $5 then null else stale_alert_at end,
        updated_at            = now()
      where id = $1
